renamed some fragments for the Progress Report Edit Form
